+++ ui/lib/controllers/main/authentication.js
@@ -1,29 +1,31 @@
 var Hapi = require('hapi');
+var Boom = require('boom');
 var debug = require('debug')('frontend:authentication');
 
 // Rutha says: You should AJAX authentication calls
 exports.login = function(request, reply) {
 
-    var pack = request.server.pack;
-    var UserModel = pack.app.mongoose.models.User;
+    var server = request.server;
+    var UserModel = server.app.mongoose.models.User;
 
     UserModel.findOne({ email: request.payload.email }, function(err, user) {
 
       if (err) {
         reply().redirect('/#login');
-        pack.app.logger.error(err);
+        server.app.logger.error(err);
         return;
       }
 
       if (!user) {
-        reply(Hapi.error.notFound('User not found'));
+        reply(Boom.notFound('User not found'));
+        return;
       }
 
       if (user && user.validPassword(request.payload.password)) {
         request.auth.session.set(user);
         reply().redirect('/profile');
       } else {
-        reply(Hapi.error.badRequest('Invalid password'));
+        reply(Boom.badRequest('Invalid password'));
       }
 
     });
@@ -36,24 +38,24 @@ exports.logout = function(request, reply) {
  };
 
 exports.signup = function(request, reply) {
-      var pack = request.server.pack;
-      var UserModel = pack.app.mongoose.models.User;
+      var server = request.server;
+      var UserModel = server.app.mongoose.models.User;
 
       UserModel.signup({
         email: request.payload.email,
         password: request.payload.password
       }, function(err, model) {
         if (err) {
-          var internalErr = Hapi.error.internal('Failed signup user', err);
+          var internalErr = Boom.internal('Failed signup user', err);
           reply(internalErr);
-          pack.app.logger.error(err);
+          server.app.logger.error(err);
           return;
         }
 
         if (model) {
           reply().redirect('/#login');
         } else {
-          reply(Hapi.error.badRequest('Username already exists'));
+          reply(Boom.badRequest('Username already exists'));
         }
       });
 };